Revert broader refactoring.
