+++ core/src/main/scala/org/apache/spark/rdd/SubtractedRDD.scala
@@ -116,7 +116,8 @@ private[spark] class SubtractedRDD[K: ClassTag, V: ClassTag, W: ClassTag](
 
         case shuffleDependency: ShuffleDependency[_, _, _] =>
           val iter = SparkEnv.get.shuffleManager
-            .getReader(shuffleDependency.shuffleHandle, partition.index, context)
+            .getReader(
+              shuffleDependency.shuffleHandle, partition.index, partition.index + 1, context)
             .read()
           iter.foreach(op)
       }
