+++ core/src/test/scala/org/apache/spark/scheduler/AdaptiveSchedulingSuite.scala
@@ -34,7 +34,7 @@ class CustomShuffledRDD[K, V, C](@transient dep: ShuffleDependency[K, V, C])
 
   override def compute(split: Partition, context: TaskContext): Iterator[(K, C)] = {
     val dep = dependencies.head.asInstanceOf[ShuffleDependency[K, V, C]]
-    SparkEnv.get.shuffleManager.getReader(dep.shuffleHandle, split.index, context)
+    SparkEnv.get.shuffleManager.getReader(dep.shuffleHandle, split.index, split.index + 1, context)
       .read()
       .asInstanceOf[Iterator[(K, C)]]
   }
