+++ core/src/test/scala/org/apache/spark/shuffle/hash/HashShuffleReaderSuite.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.shuffle
+package org.apache.spark.shuffle.hash
 
 import java.io.{ByteArrayOutputStream, InputStream}
 import java.nio.ByteBuffer
@@ -28,6 +28,7 @@ import org.mockito.stubbing.Answer
 import org.apache.spark._
 import org.apache.spark.network.buffer.{ManagedBuffer, NioManagedBuffer}
 import org.apache.spark.serializer.JavaSerializer
+import org.apache.spark.shuffle.BaseShuffleHandle
 import org.apache.spark.storage.{BlockManager, BlockManagerId, ShuffleBlockId}
 
 /**
@@ -55,10 +56,10 @@ class RecordingManagedBuffer(underlyingBuffer: NioManagedBuffer) extends Managed
   }
 }
 
-class ShuffleReaderSuite extends SparkFunSuite with LocalSparkContext {
+class HashShuffleReaderSuite extends SparkFunSuite with LocalSparkContext {
 
   /**
-   * This test makes sure that, when data is read from a ShuffleReader, the underlying
+   * This test makes sure that, when data is read from a HashShuffleReader, the underlying
    * ManagedBuffers that contain the data are eventually released.
    */
   test("read() releases resources on completion") {
@@ -124,7 +125,7 @@ class ShuffleReaderSuite extends SparkFunSuite with LocalSparkContext {
       Seq((localBlockManagerId, shuffleBlockIdsAndSizes))
     }
 
-    // Create a mocked shuffle handle to pass into ShuffleReader.
+    // Create a mocked shuffle handle to pass into HashShuffleReader.
     val shuffleHandle = {
       val dependency = mock(classOf[ShuffleDependency[Int, Int, Int]])
       when(dependency.serializer).thenReturn(Some(serializer))
@@ -133,9 +134,10 @@ class ShuffleReaderSuite extends SparkFunSuite with LocalSparkContext {
       new BaseShuffleHandle(shuffleId, numMaps, dependency)
     }
 
-    val shuffleReader = new ShuffleReader(
+    val shuffleReader = new HashShuffleReader(
       shuffleHandle,
       reduceId,
+      reduceId + 1,
       TaskContext.empty(),
       blockManager,
       mapOutputTracker)
